rename layers to base_archs
