+++ mmedit/models/editors/basicvsr/basicvsr.py
@@ -55,7 +55,7 @@ class BasicVSR(BaseEditModel):
         self.forward_ensemble = None
         if ensemble is not None:
             if ensemble['type'] == 'SpatialTemporalEnsemble':
-                from mmedit.models.layers import SpatialTemporalEnsemble
+                from mmedit.models.base_archs import SpatialTemporalEnsemble
                 is_temporal = ensemble.get('is_temporal_ensemble', False)
                 self.forward_ensemble = SpatialTemporalEnsemble(is_temporal)
             else:
