+++ mmedit/models/editors/glean/glean_styleganv2.py
@@ -4,7 +4,7 @@ import torch
 import torch.nn as nn
 from mmengine.model import BaseModule
 
-from mmedit.models.layers import PixelShufflePack
+from mmedit.models.base_archs import PixelShufflePack
 from mmedit.models.utils import make_layer
 from mmedit.registry import COMPONENTS, MODELS
 from ..esrgan.rrdb_net import RRDB
