+++ mmedit/models/editors/srgan/sr_resnet.py
@@ -2,7 +2,7 @@
 import torch.nn as nn
 from mmengine.model import BaseModule
 
-from mmedit.models.layers import PixelShufflePack, ResidualBlockNoBN
+from mmedit.models.base_archs import PixelShufflePack, ResidualBlockNoBN
 from mmedit.models.utils import default_init_weights, make_layer
 from mmedit.registry import MODELS
