+++ mmedit/models/editors/ttsr/lte.py
@@ -3,7 +3,7 @@ import torch
 from mmengine.model import BaseModule
 from torchvision import models
 
-from mmedit.models.layers import ImgNormalize
+from mmedit.models.base_archs import ImgNormalize
 from mmedit.registry import MODELS
