+++ mmedit/models/editors/ttsr/ttsr_net.py
@@ -6,7 +6,7 @@ import torch.nn.functional as F
 from mmcv.cnn import build_conv_layer
 from mmengine.model import BaseModule
 
-from mmedit.models.layers import PixelShufflePack, ResidualBlockNoBN
+from mmedit.models.base_archs import PixelShufflePack, ResidualBlockNoBN
 from mmedit.models.utils import make_layer
 from mmedit.registry import MODELS
