+++ tests/test_models/test_base_archs/test_conv.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from mmedit.models.layers import conv
+from mmedit.models.base_archs import conv
 
 
 def test_conv():
