+++ tests/test_models/test_base_archs/test_gated_conv_module.py
@@ -3,7 +3,7 @@ import pytest
 import torch
 import torch.nn as nn
 
-from mmedit.models.layers import SimpleGatedConvModule
+from mmedit.models.base_archs import SimpleGatedConvModule
 
 
 def test_gated_conv():
