+++ tests/test_models/test_base_archs/test_img_normalize.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import torch
 
-from mmedit.models.layers import ImgNormalize
+from mmedit.models.base_archs import ImgNormalize
 
 
 def test_normalize_layer():
