+++ tests/test_models/test_base_archs/test_upsample.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import torch
 
-from mmedit.models.layers import PixelShufflePack
+from mmedit.models.base_archs import PixelShufflePack
 
 
 def test_pixel_shuffle():
