+++ tests/test_models/test_editors/test_deepfillv1/test_contextual_attention_neck.py
@@ -1,8 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import torch
 
+from mmedit.models.base_archs import SimpleGatedConvModule
 from mmedit.models.editors import ContextualAttentionNeck
-from mmedit.models.layers import SimpleGatedConvModule
 
 
 def test_deepfill_contextual_attention_neck():
