+++ tests/test_models/test_editors/test_global_local/test_gl_dilation.py
@@ -1,8 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import torch
 
+from mmedit.models.base_archs import SimpleGatedConvModule
 from mmedit.models.editors.global_local import GLDilationNeck
-from mmedit.models.layers import SimpleGatedConvModule
 from mmedit.registry import MODELS
 from mmedit.utils import register_all_modules
